refactor: split utility into separate files
