+++ src/utils/compareItems.ts
@@ -1,0 +1,11 @@
+import { SortCallback } from "../types";
+
+export const compareItems: SortCallback = (a, b) => {
+  if (a.text > b.text) {
+    return 1;
+  } else if (a.text < b.text) {
+    return -1;
+  }
+
+  return 0;
+};
